requirements_base.txt
aiohttp==3.9.4
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
elastic-transport==8.13.0
elasticsearch==8.13.0
environs==9.5.0
et-xmlfile==1.1.0
flask==3.0.3
frozenlist==1.4.1
grpcio==1.60.0
idna==3.7
importlib-metadata==7.1.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
minio==7.2.5
multidict==6.0.5
mysql-connector-python==8.3.0
numpy==1.24.4
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
protobuf==5.26.1
psutil==5.9.8
pyarrow==15.0.2
pyBrainai==1.4.1.0
pycparser==2.22
pycryptodome==3.20.0
pymilvus==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
schedule==1.2.1
scipy==1.10.1
six==1.16.0
tiktoken==0.6.0
typing-extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
werkzeug==3.0.2
yarl==1.9.4
zipp==3.18.1
